Set up tests for normal SM logic
In order to allow for testing the logic for some location, all prizes (except Aghanim) are removed to take advantage of the fact an empty collection satisfies an "all" condition. To make this work universally, `CanAquireAll` takes the place of `CanAquire` so that the condition is always applied to a collection.